remove controller and dto and restructure
