+++ packages/backend/src/db/db.module.ts
@@ -1,14 +1,12 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { SampleController } from './db.controller';
-import { SampleService } from './db.service';
-import { Sample, SampleSchema } from './schemas/sample.schema';
+import { SampleService } from './sample/sample.service';
+import { Sample, SampleSchema } from './sample/sample.schema';
 
 @Module({
   imports: [
     MongooseModule.forFeature([{ name: Sample.name, schema: SampleSchema }]),
   ],
-  controllers: [SampleController],
   providers: [SampleService],
 })
 export class SampleModule {}
