--- packages/backend/src/db/dto/create-sample.dto.ts
@@ -1,5 +1,0 @@
-export class CreateSampleDto {
-  readonly name: string;
-  readonly age: number;
-  readonly breed: string;
-}
